requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
